These are the (hopefully complete, I will add if it turns out I missed something) analysis and figure scripts for our paper Comparing Data-Driven Subtypes of Depression!
Sorry if the organization (and the commenting) of these scripts are terribly confusing. We wanted to have something up by the time the paper went live but they're not in the state that I hoped they would be in by this point (you know, understandable to anyone but me...). These are also only the scripts related to analysis and figures, and not data wrangling. This was to be sure we were being absolutely compliant with UKB's data sharing policies. But if udnerstanding how the data is structured is helpful, let us know and we can create versions of those scripts that should be good for sharing.